requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
jinja2
starlette-session
